pathing almost there

+++ src/Generator/Layout.hpp
@@ -49,8 +49,8 @@ namespace SG::Generator
             ||  this->assets.find( TILE::WALL ) == this->assets.end() )
             { throw "Not enough assets to generate layout"; }
 
-            int yS = layout->getSize().y;
-            int xS = layout->getSize().x;
+            uint16_t yS = layout->getSize().y;
+            uint16_t xS = layout->getSize().x;
 
             for( size_t y=0; y < yS; y++)
             {
